Isolate generation planning by use case
